+++ src/controllers/utils/list.messages.ts
@@ -1,16 +1,19 @@
 enum ErrorMessagesForLists {
   alreadyFinished = 'La lista ya finalizó su inscripción.',
+  differentFaculty = 'Usted no puede inscribir una lista en una facultad que no es la suya.',
   errorCreating = 'Hubo un error creando la lista.',
   errorEnrolling = 'Hubo un problema inscribiendo al ',
   errorFiltering = 'Hubo un problema obteniendo las listas solicitadas.',
   errorFinishingRegistration = 'Hubo un problema finalizando la inscripción de su lista',
   errorGettingLists = 'Hubo un error tratando de obtener las listas registradas.',
-  limitList = 'Usted no puede registrar más de 2 listas.',
+  limitList = 'Usted no puede inscribir más de 2 listas.',
   missingList = 'La lista solicitada no existe.',
   missingOwner = 'El dueño de la lista no está registrado.',
   noList = ' no tiene ninguna lista registrada.',
+  studentListAlready = 'Usted ya no puede inscribir otra lista de estudiantes.',
+  teacherListAlready = 'Usted ya no puede inscribir otra lista de docentes.',
   unauthorizedFinish = 'Usted no puede finalizar la inscripción de una lista que no representa.',
-  unauthorizedRegistration = 'Usted no puede registrar una lista porque no está registrado.'
+  unauthorizedRegistration = 'Usted no puede inscribir una lista porque no está registrado.'
 }
 
 enum MessagesForLists {